Reorganise benchmarks and tests for future
