+++ Interpreter/src/bench/java/org/enso/interpreter/bench/BenchmarkItem.java
@@ -1,4 +1,4 @@
-package org.enso.interpreter;
+package org.enso.interpreter.bench;
 
 import org.openjdk.jmh.results.Result;
