+++ Interpreter/src/bench/java/org/enso/interpreter/bench/BenchmarksRunner.java
@@ -1,4 +1,4 @@
-package org.enso.interpreter;
+package org.enso.interpreter.bench;
 
 import org.openjdk.jmh.results.RunResult;
 import org.openjdk.jmh.runner.BenchmarkList;
