+++ Interpreter/src/bench/java/org/enso/interpreter/bench/ReportItem.java
@@ -1,4 +1,4 @@
-package org.enso.interpreter;
+package org.enso.interpreter.bench;
 
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlElementWrapper;
