+++ Interpreter/src/bench/java/org/enso/interpreter/bench/benchmarks/semantic/RecursionBenchmarks.java
@@ -1,6 +1,6 @@
-package org.enso.interpreter.benchmarks;
+package org.enso.interpreter.bench.benchmarks.semantic;
 
-import org.enso.interpreter.fixtures.RecursionFixtures;
+import org.enso.interpreter.bench.fixtures.semantic.RecursionFixtures;
 import org.openjdk.jmh.annotations.*;
 
 import java.util.concurrent.TimeUnit;
