+++ Interpreter/src/bench/scala/org/enso/interpreter/bench/RegressionTest.scala
@@ -1,4 +1,4 @@
-package org.enso.interpreter
+package org.enso.interpreter.bench
 
 import org.scalatest.{FlatSpec, Matchers}
