+++ Interpreter/src/bench/scala/org/enso/interpreter/bench/fixtures/semantic/RecursionFixtures.scala
@@ -1,6 +1,7 @@
-package org.enso.interpreter.fixtures
+package org.enso.interpreter.bench.fixtures.semantic
 
 import org.enso.interpreter.Constants
+import org.enso.interpreter.test.semantic.LanguageRunner
 
 class RecursionFixtures extends LanguageRunner {
   val hundredMillion: Long = 100000000
