+++ Interpreter/src/test/scala/org/enso/interpreter/test/semantic/MethodsTest.scala
@@ -1,4 +1,4 @@
-package org.enso.interpreter.semantic
+package org.enso.interpreter.test.semantic
 
 import org.graalvm.polyglot.PolyglotException
